Adding a client file to go along with server file.
Chaging default port to 6563 (teehee).
